added delete and admin functionality

+++ src/commands/user/user.controller.ts
@@ -2,6 +2,7 @@ import { CommandBus } from '@nestjs/cqrs';
 import { JwtAuthGuard } from '../../auth/jwt-auth.guard';
 import { DeleteUserBody } from './models/bodies/delete-body';
 import { DeleteUserCommand } from './commands/delete.command';
+import { NoRightsException } from '../owner/errors/no-rights-exception';
 import { DomainExceptionFilter } from './errors/domain-exception.filter';
 import { ApiTags, ApiResponse, ApiOperation, ApiBearerAuth } from '@nestjs/swagger';
 import { UseFilters, Controller, Delete, UseGuards, Request, Body } from '@nestjs/common';
@@ -30,6 +31,10 @@ export class UserController {
     @ApiResponse({ status: 200, description: 'User removed' })
     @ApiResponse({ status: 400, description: 'User removal failed' })
     async removeOwnerById(@Request() req, @Body() body: DeleteUserBody) {
-        return await this.commandBus.execute(new DeleteUserCommand(body.email));
+        if (req.user.isAdmin || req.user.isStaff) {
+            return await this.commandBus.execute(new DeleteUserCommand(body.email));
+        } else {
+            throw new NoRightsException(req.user);
+        }
     }
 }
